refactor: rename ApizrTests and ApizrWithCancellationTokensTests files and update namespace to Refitter.Tests.Apizr
